third_party/WebKit/LayoutTests/http/tests/devtools/elements/edit/shadow-dom-modify-chardata-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that elements panel updates shadow dom tree structure upon typing.


Running: testDumpInitial
========= Original ========
- <div id="container">
    - <input type="text" id="input1">
        - #shadow-root (user-agent)
              <div></div>
      </input>
  </div>

Running: testAppend
======== Type text =========
- <div id="container">
    - <input type="text" id="input1">
        - #shadow-root (user-agent)
              <div>Bar</div>
      </input>
  </div>
